requirements.txt
boto==2.49.0
boto3==1.9.80
botocore==1.12.80
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dill==0.2.8.2
docutils==0.14
docx==0.2.4
Flask==1.0.2
gensim==3.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
lxml==4.3.0
MarkupSafe==1.1.0
murmurhash==1.0.1
nltk==3.4
numpy==1.16
pandas==0.23.4
pdfminer.six==20181108
Pillow==5.4.1
pycryptodome==3.7.2
pyreadline==2.1
python-dateutil==2.7.5
python-docx==0.8.10
pytz==2018.9
PyYAML==3.13
requests==2.21.0
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.0
sortedcontainers==2.1.0
toolz==0.9.0
tqdm==4.29.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
